# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/token": {
      "post": {
        "summary": "Login For Access Token",
        "description": "Verifies a client's login request and returns a token if it succeeds.\n\nArgs:\n    db: Database dependency.\n    form_data: OAuth2PasswordRequestForm dependency.",
        "operationId": "login_for_access_token_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register User",
        "description": "Endpoint for registering a new user. Client with registration request must not be logged in.\n\nArgs:\n    request: Create user request.\n    db: Database dependency.\n    current_user: Optional user dependency. Should be None for this function to succeed.",
        "operationId": "register_user_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/users": {
      "post": {
        "summary": "Create User",
        "description": "Creates a user with metadata request, provided that the current_user has sufficient permissions.\n\nArgs:\n    db: Database dependency.\n    current_user: Current user dependency. Must not be None (e.g. user must be logged in).\n    request: Metadata of user to be added.",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "summary": "Read User Me",
        "description": "Return the current logged in user.\n\nArgs:\n    current_user: Current user dependency.",
        "operationId": "read_user_me_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete User Me",
        "description": "Delete the user currently logged in.",
        "operationId": "delete_user_me_users_me_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/users/{userName}": {
      "get": {
        "summary": "Read User",
        "description": "Get user by username.\n\nArgs:\n    user_name: Username of user.\n    db: Database dependency.",
        "operationId": "read_user_users__userName__get",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "delete": {
        "summary": "Delete User",
        "description": "Delete the user with user_id if the current user has sufficient permissions to perform this action. Throw an exception if the user currently logged in has insufficient permissions.",
        "operationId": "delete_user_users__userId__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Userid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auto-labels/{autoLabelId}": {
      "get": {
        "summary": "Read Autolabel By Id",
        "description": "Endpoint for retrieving autolabel from database.\n\nArgs:\n    auto_label_id: UUID for auto label.\n    db: Database dependency.\n    current_user: Current user dependency.",
        "operationId": "read_autolabel_by_id_auto_labels__autoLabelId__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "autoLabelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Autolabelid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoLabel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auto-labels": {
      "get": {
        "summary": "Read Autolabels",
        "operationId": "read_autolabels_auto_labels_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "novelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Novelid"
            }
          },
          {
            "name": "chapterIds",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapterids"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "modelNames",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modelnames"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutoLabelMeta"
                  },
                  "title": "Response Read Autolabels Auto Labels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Autolabels",
        "operationId": "create_autolabels_auto_labels_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAutoLabels"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutoLabelMeta"
                  },
                  "title": "Response Create Autolabels Auto Labels Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoLabel": {
        "properties": {
          "autoLabelId": {
            "type": "string",
            "format": "uuid",
            "title": "Autolabelid"
          },
          "autoLabelData": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LabelBase"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Autolabeldata"
          },
          "autoLabelModelName": {
            "type": "string",
            "title": "Autolabelmodelname"
          },
          "autoLabelModelParams": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 50,
            "title": "Autolabelmodelparams"
          },
          "autoLabelStatus": {
            "$ref": "#/components/schemas/AutoLabelProgress"
          },
          "autoLabelMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Autolabelmessage"
          },
          "chapterContentId": {
            "type": "string",
            "format": "uuid",
            "title": "Chaptercontentid"
          },
          "autoLabelLastJobId": {
            "type": "string",
            "title": "Autolabellastjobid"
          }
        },
        "type": "object",
        "required": [
          "autoLabelId",
          "autoLabelData",
          "autoLabelModelName",
          "autoLabelModelParams",
          "autoLabelStatus",
          "chapterContentId",
          "autoLabelLastJobId"
        ],
        "title": "AutoLabel",
        "description": "Pydantic schema for an auto-labeled data entry.\n\nAttributes:\n    auto_label_id: UUID identifier for this AutoLabel.\n    auto_label_data: Dictionary containing the auto-labeled data.\n    auto_label_model_name: Name of the model used to generate the auto labels.\n    auto_label_model_params: Parameters used for the model to generate the auto labels.\n    auto_label_status: Labeling progress for this autolabel.\n    auto_label_message: Details on status.\n    chapter_content_id: UUID of chapter content this AutoLabel is associated with.\n    auto_label_last_job_id: Job id of last job that was run on this AutoLabel."
      },
      "AutoLabelMeta": {
        "properties": {
          "autoLabelId": {
            "type": "string",
            "format": "uuid",
            "title": "Autolabelid"
          },
          "autoLabelModelName": {
            "type": "string",
            "title": "Autolabelmodelname"
          },
          "autoLabelModelParams": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 50,
            "title": "Autolabelmodelparams"
          },
          "autoLabelStatus": {
            "$ref": "#/components/schemas/AutoLabelProgress"
          },
          "autoLabelMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Autolabelmessage"
          },
          "chapterContentId": {
            "type": "string",
            "format": "uuid",
            "title": "Chaptercontentid"
          },
          "autoLabelLastJobId": {
            "type": "string",
            "title": "Autolabellastjobid"
          }
        },
        "type": "object",
        "required": [
          "autoLabelId",
          "autoLabelModelName",
          "autoLabelModelParams",
          "autoLabelStatus",
          "chapterContentId",
          "autoLabelLastJobId"
        ],
        "title": "AutoLabelMeta",
        "description": "Pydantic schema for auto-label metadata.\n\nAttributes:\n    auto_label_id: UUID identifier for this AutoLabel.\n    auto_label_model_name: Name of the model used to generate the auto labels.\n    auto_label_model_params: Parameters used for the model to generate the auto labels.\n    auto_label_status: Labeling progress for this autolabel.\n    auto_label_message: Details on status.\n    chapter_content_id: UUID of chapter content this AutoLabel is associated with.\n    auto_label_last_job_id: Job id of last job that was run on this AutoLabel."
      },
...
      "Body_login_for_access_token_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_for_access_token_token_post"
      },
...
      "CreateAutoLabels": {
        "properties": {
          "novelId": {
            "type": "string",
            "format": "uuid",
            "title": "Novelid"
          },
          "autoLabelModelName": {
            "type": "string",
            "title": "Autolabelmodelname"
          },
          "autoLabelModelParams": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 50,
            "title": "Autolabelmodelparams"
          },
          "chapterIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapterids"
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "isPublic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ispublic"
          }
        },
        "type": "object",
        "required": [
          "novelId",
          "autoLabelModelName",
          "autoLabelModelParams"
        ],
        "title": "CreateAutoLabels",
        "description": "Pydantic schema for creating an auto-labeled data entry.\n\nAttributes:\n    novel_id: UUID of novel to create auto labels for.\n    auto_label_model_name: Name of the model used to generate the auto labels.\n    auto_label_model_params: Parameters used for the model to generate the auto labels.\n    chapter_ids: Optional parameter. Restrict to revisions with specific chapter UUIDs.\n    start: Optional parameter. Restrict to revisions with chapter num >= start.\n    end: Optional parameter. Restrict to revisions with chapter num < end.\n    is_public: Optional parameter. Restrict to revisions with this specific public flag."
      },
...
      "CreateUser": {
        "properties": {
          "userName": {
            "type": "string",
            "title": "Username"
          },
          "userPassword": {
            "type": "string",
            "title": "Userpassword"
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          }
        },
        "type": "object",
        "required": [
          "userName",
          "userPassword",
          "userType"
        ],
        "title": "CreateUser",
        "description": "Pydantic schema to validate data required to create a user.\n\nAttributes:\n    user_name: Username of user being created.\n    user_password: Unhashed password of user being create.\n    user_type: UserType type of user to be created.\nNotes:"
      },
...
      "DeleteUserStatus": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "fail",
              "verify"
            ],
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DeleteUserStatus",
        "description": "Pydantic schema to return status after deleting a user\n\nAttributes:\n    status: One of 'success', 'fail', 'verify'.\n    detail: String denoting details of operation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token",
        "description": "Pydantic schema for a JWT\n\nAttributes:\n    access_token: Payload.\n    token_type: Type of token."
      },
...
      "User": {
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "userName": {
            "type": "string",
            "title": "Username"
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "userName",
          "userType"
        ],
        "title": "User",
        "description": "Pydantic model for user information.\n\nAttributes:\n    user_id: id of this user.\n    user_name: Username for this user.\n    user_type: A value in UserType (e.g. 'admin', or 'user')"
      },
      "UserType": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ],
        "title": "UserType",
        "description": "Enum of possible user types."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}